+++ packages/analytics/analytics-chart/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.3.1...@kong-ui-public/analytics-chart@2.3.2) (2024-03-04)
+
+
+### Bug Fixes
+
+* adjust one more time ([#1205](https://github.com/Kong/public-ui-components/issues/1205)) ([f7d6dd8](https://github.com/Kong/public-ui-components/commit/f7d6dd810da2eaa75fd73463dcddef5f09f01e6e))
+
+
+
+
+
+## [2.3.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.3.0...@kong-ui-public/analytics-chart@2.3.1) (2024-03-04)
+
+
+### Bug Fixes
+
+* adjust legend height ([#1204](https://github.com/Kong/public-ui-components/issues/1204)) ([da758e0](https://github.com/Kong/public-ui-components/commit/da758e0688183e170612fd3ab46b7f31d1ccf07a))
+
+
+
+
+
+# [2.3.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.2.5...@kong-ui-public/analytics-chart@2.3.0) (2024-03-01)
+
+
+### Features
+
+* **analytics-chart:** title and legend [MA-2670] ([#1198](https://github.com/Kong/public-ui-components/issues/1198)) ([d6766dd](https://github.com/Kong/public-ui-components/commit/d6766dd0b017f7bde67309c4265df52c928ec317))
+
+
+
+
+
 ## [2.2.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.2.4...@kong-ui-public/analytics-chart@2.2.5) (2024-02-27)